Resolve invalid syntax error when attempting to purge empty array of IDs.
git-svn-id: http://plugins.svn.wordpress.org/document-gallery@1314447 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ trunk/inc/class-thumb.php
@@ -367,7 +367,7 @@ class DG_Thumb {
 
                 self::$thumbs = null;
             }
-        } else {
+        } else if ( !empty( $ids ) ) {
             $sql = "DELETE FROM $postmeta WHERE meta_key = '$meta_key' AND post_id IN(" . rtrim( str_repeat( '%d,', sizeof( $ids ) ), ',' ) . ")";
             $result = $wpdb->query( $wpdb->prepare( $sql, $ids ) );
